Add a test and optimize setup.
The loading of the subway datas takes 14 seconds !!!!
One can say it's not optimized. Thus we added a global variable
to load these datas only once.

Add a test to ensure that no station has less than one neighbor
nor more than three (because that's how the Paris' subway is made).
This will be false when we will add transfert by walk.